% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggnet.R
\name{ggnet}
\alias{ggnet}
\title{Draw pretty network graphs with ggplot}
\usage{
ggnet(g, node = list(label = NA, color = "grey50", size = 1),
  label = list(size = 1, color = "black"))
}
\arguments{
\item{g}{igraph object}

\item{node}{list}

\item{label}{list}
}
\description{
Draw pretty network graphs with ggplot
}
