% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maprofiler.R
\name{maprofiler}
\alias{maprofiler}
\title{Automated Microarray Profiler}
\usage{
maprofiler(setup)
}
\arguments{
\item{setup}{path to the .yaml file outlining project set up}
}
\description{
Automated Microarray Profiler
}
