% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringdb.R
\name{stringdb}
\alias{stringdb}
\title{Simple interface to STRINGdb}
\usage{
stringdb(ids, score = 900, additional_network_nodes = 1)
}
\arguments{
\item{ids}{character vector}

\item{score}{numeric}

\item{additional_network_nodes}{numeric}
}
\description{
Simple interface to STRINGdb
}
