% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoplot.R
\name{volcanoplot}
\alias{volcanoplot}
\title{Draw a labeled volcano plot}
\usage{
volcanoplot(fc, pval, lbl = NULL, ntop = length(lbl))
}
\arguments{
\item{fc}{numeric vector}

\item{pval}{numeric vector}

\item{lbl}{character vector}

\item{ntop}{numeric number of top labels to show}
}
\description{
Draw a labeled volcano plot
}
